updated to use dnvm
